+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 14 April 2022
+* Add CSS inlining into html file using webpack plugin - only in production mode
+* Small refactor of webpack config
+
 ## 28 March 2022
 * Change build directory from `app` to `dist`
 * Do not minify css in `dev` mode